--- app/assets/javascripts/actions/index.js
@@ -1,1 +1,0 @@
-
