+++ app/assets/javascripts/actions/static.js
@@ -2,7 +2,6 @@ import React from 'react';
 
 export const SET_COMMUNITY_DATA = 'SET_COMMUNITY_DATA';
 export const SET_USERS_ACTIVITY_DATA = 'SET_USERS_ACTIVITY_DATA';
-export const SET_CAMPAINGS_LIST = 'SET_CAMPAINGS_LIST';
 export const SET_EVENTS_LIST = 'SET_EVENTS_LIST';
 
 export function setCommunityData(data) {
@@ -38,19 +37,6 @@ export function setUsersActivityData(data) {
       {name: 'Pepi', roads: 33},
       {name: 'Pepi', roads: 33}
     ]
-  };
-}
-
-export function setCampaingsList() {
-  const url = '/api/v1/campaigns';
-
-   return function(dispatch) {
-    $.get(url).then(function(campaingsList){
-      dispatch({
-        type: SET_CAMPAINGS_LIST,
-        campaingsList
-      });
-    });
   };
 }
