+++ app/assets/javascripts/reducers/index.js
@@ -1,6 +1,7 @@
-
 import staticReducer from './static';
+import campaingsReducer from './campaingsReducer';
 
 export default {
+  campaingsReducer,
   staticReducer
 };
