+++ app/assets/javascripts/reducers/static.js
@@ -8,10 +8,6 @@ function setUsersActivityData(state, usersActivityData) {
   return Object.assign({}, state, {usersActivityData});
 }
 
-function setCampaingsList(state, campaingsList) {
-  return Object.assign({}, state, {campaingsList});
-}
-
 function setEventsList(state, eventsList) {
   return Object.assign({}, state, {eventsList});
 }
@@ -22,8 +18,6 @@ export default function staticReducer(state = {} , action) {
       return setCommunityData(state, action.communityData);
     case 'SET_USERS_ACTIVITY_DATA':
       return setUsersActivityData(state, action.usersActivityData);
-    case 'SET_CAMPAINGS_LIST':
-      return setCampaingsList(state, action.campaingsList);
     case 'SET_EVENTS_LIST':
       return setEventsList(state, action.eventsList);
     default:
